backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
scikit-learn
xgboost
pandas
numpy
google-generativeai
